pyproject.toml
[project]
name = "doc-gen-ai-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "box-sdk-gen>=1.10.0",
    "dataclasses-json>=0.6.7",
    "dataclasses-jsonschema>=2.16.0",
    "python-dotenv>=1.0.1",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "coverage>=7.6.10",
    "pytest>=8.3.4",
    "ruff>=0.9.3",
]
